ref: maked app/auth path and splitted layout
